Relax Chain ID format requirement for EIP-712 verification
The upstream Eip712SigVerificationDecorator requires the Cosmos Chain ID
to adhere to a format like "evmos_7001-1" which does not work for
Gravity's "gravity-bridge-3" format. This commit adds a EvmChainId value
to the AnteHandler HandlerOptions so that Gravity can override that
check and use its existing Chain ID format.

Chain IDs are difficult to change and cause a lot of issues with IBC
channels once changed. By relaxing this restriction the integration of
EIP-712 is made much simpler.